separate vulkan fmt formatters to vk_fmt_formatters.h
